Crate ipld_amt

AMT crate for use as rust IPLD data structure

Data structure reference: https://github.com/ipld/specs/blob/51fab05b4fe4930d3d851d50cc1e5f1a02092deb/data-structures/vector.md

Structs

Amt
Array Mapped Trie allows for the insertion and persistence of data, serializable to a CID.

ValueMut
A wrapper around a mutable reference to a value in an Amt. Keeps track of whether the value was mutated by setting a flag when DerefMut::deref_mut is called.

Enums

Error
AMT Error

Constants

MAX_INDEX
MaxIndex is the maximum index for elements in the AMT. This u64::MAX-1 so we don’t overflow u64::MAX when computing the length.