Struct ipld_amt::AMT

pub struct AMT<'db, DB, V> where
    DB: BlockStore,
    V: Clone,  { /* fields omitted */ }

Array Mapped Trie allows for the insertion and persistence of data, serializable to a CID

Usage:

use ipld_amt::AMT;

let db = db::MemoryDB::default();
let mut amt = AMT::new(&db);

// Insert or remove any serializable values
amt.set(2, "foo".to_owned()).unwrap();
amt.set(1, "bar".to_owned()).unwrap();
amt.delete(2).unwrap();
assert_eq!(amt.count(), 1);
let bar: String = amt.get(1).unwrap().unwrap();

// Generate cid by calling flush to remove cache
let cid = amt.flush().unwrap();

Methods

impl<'db, DB, V> AMT<'db, DB, V> where
DB: BlockStore,
V: Clone + DeserializeOwned + Serialize,

pub fn new(block_store: &'db DB) -> Self

Constructor for Root AMT node

pub fn load(block_store: &'db DB, cid: &Cid) -> Result<Self, Error>

Constructs an AMT with a blockstore and a Cid of the root of the AMT

pub fn height(&self) -> u32

pub fn count(&self) -> u64

pub fn new_from_slice(block_store: &'db DB, vals: &[V]) -> Result<Cid, Error>

Generates an AMT with block store and array of cbor marshallable objects and returns Cid

pub fn get(&self, i: u64) -> Result<Option<V>, Error>

Get value at index of AMT

pub fn set(&mut self, i: u64, val: V) -> Result<(), Error>

Set value at index

pub fn batch_set(&mut self, vals: &[V]) -> Result<(), Error>

Batch set (naive for now)

pub fn delete(&mut self, i: u64) -> Result<bool, Error>

Delete item from AMT at index

pub fn flush(&mut self) -> Result<Cid, Error>

flush root and return Cid used as key in block store